move logger pkg to root dir
