+++ agent/agent.go
@@ -17,7 +17,7 @@ import (
 	"github.com/netdata/go.d.plugin/agent/job/state"
 	"github.com/netdata/go.d.plugin/agent/module"
 	"github.com/netdata/go.d.plugin/agent/netdataapi"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 	"github.com/netdata/go.d.plugin/pkg/multipath"
 
 	"github.com/mattn/go-isatty"
