+++ agent/job/discovery/dummy/discovery.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 
 	"github.com/netdata/go.d.plugin/agent/job/confgroup"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 type Config struct {
