+++ agent/job/discovery/file/discovery.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 
 	"github.com/netdata/go.d.plugin/agent/job/confgroup"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 type Config struct {
