+++ agent/job/discovery/file/watch.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/netdata/go.d.plugin/agent/job/confgroup"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 
 	"github.com/fsnotify/fsnotify"
 )
