+++ agent/job/discovery/manager.go
@@ -10,7 +10,7 @@ import (
 	"github.com/netdata/go.d.plugin/agent/job/confgroup"
 	"github.com/netdata/go.d.plugin/agent/job/discovery/dummy"
 	"github.com/netdata/go.d.plugin/agent/job/discovery/file"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 type Config struct {
