+++ agent/job/run/run.go
@@ -7,7 +7,7 @@ import (
 
 	jobpkg "github.com/netdata/go.d.plugin/agent/job"
 	"github.com/netdata/go.d.plugin/agent/ticker"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 type (
