+++ agent/job/state/state.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/netdata/go.d.plugin/agent/job/confgroup"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 type Manager struct {
