+++ agent/module/job.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/netdata/go.d.plugin/agent/netdataapi"
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 )
 
 var writeLock = &sync.Mutex{}
