+++ logger/countwatcher_test.go
