+++ logger/severity.go
