+++ logger/static.go
