+++ modules/vsphere/scrape/scrape.go
@@ -9,7 +9,7 @@ import (
 
 	rs "github.com/netdata/go.d.plugin/modules/vsphere/resources"
 
-	"github.com/netdata/go.d.plugin/pkg/logger"
+	"github.com/netdata/go.d.plugin/logger"
 	"github.com/vmware/govmomi/performance"
 	"github.com/vmware/govmomi/vim25/types"
 )
